Move the test data sets to their own internal file.
The dataclasses were getting complex and the number of tests was growing, so it took up a large portion of test_regions.py.

Also, this way it can theoretically be reused for other tests in the future.
